+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.2.9
+
+- More Excalidraw tools support.
+  - `Frame tool`, `Web Embed`, `Text to diagram`, `Mermaid to Excalidraw`, `Wireframe to code`.
+- Lossless PDF Export.
+- Fallback to `Excalifont`.
+- Optimize Excalidraw data compatibility.
+
 ## 1.2.8
 
 * Optimize i18n, now supports `English`, `简体中文`, `繁體中文`, `日本語`, `한국어`, `Español`, `Français`, `Deutsch`, `Italiano`, `Русский`, `Português`, `Nederlands`, `Polski`, `Türkçe`, `العربية`, `हिन्दी`, `ไทย`, `Tiếng Việt`